# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  uid: string;
  displayName: string;
  email: string;
  photoURL?: string;
  createdAt: string;
  updatedAt: string;
  lastActiveAt: string;
}

export interface UpdateUserProfile {
  displayName?: string;
  email?: string;
  photoURL?: string;
}

export interface UserPreferences {
  id: string;
  uid: string;
  defaultTransport?: string;
  defaultBudget?: number;
  defaultTime?: number;
  dietaryRestrictions: string[];
  accessibilityNeeds: string[];
  favoriteCategories: string[];
  notificationsEnabled: boolean;
  emailUpdates: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateUserPreferences {
  defaultTransport?: string;
  defaultBudget?: number;
  defaultTime?: number;
  dietaryRestrictions?: string[];
  accessibilityNeeds?: string[];
  favoriteCategories?: string[];
  notificationsEnabled?: boolean;
  emailUpdates?: boolean;
}

export interface ErrorResponse {
  error: string;
}