# lib/api-zod/src/generated/types/updateUserPreferences.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUserPreferences {
  defaultTransport?: string;
  defaultBudget?: number;
  defaultTime?: number;
  dietaryRestrictions?: string[];
  accessibilityNeeds?: string[];
  favoriteCategories?: string[];
  notificationsEnabled?: boolean;
  emailUpdates?: boolean;
}