# lib/api-zod/src/generated/types/updateUserProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUserProfile {
  displayName?: string;
  email?: string;
  photoURL?: string;
}