# lib/api-zod/src/generated/types/userPreferences.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPreferences {
  id: string;
  uid: string;
  defaultTransport?: string;
  defaultBudget?: number;
  defaultTime?: number;
  dietaryRestrictions: string[];
  accessibilityNeeds: string[];
  favoriteCategories: string[];
  notificationsEnabled: boolean;
  emailUpdates: boolean;
  createdAt: string;
  updatedAt: string;
}